Major rewrite to Jeepyb Scripts
This commit is a huge rewrite to the core Jeepyb functionality.
The bulk of the changes was to clean up the large number of "global"
shared settings being passed around to all functions. This change
creates a few classes to hold these settings and handle most of the
common tasks.

In addition, this change also extends the manage_projects by adding
support for creating per project Groups and Prolog Rules.