Sync tags when tracking upstream repos
The tags from the upstream repo were not always tracked. This
commit makes sure they get updated in the sync. In addition, makes
the upstream remote use the same SSH key as the Gerrit project
creater. It is up to the user to make sure the remote has an
identical user with SSH created.